requirements.txt
requests==2.28.1
validators==0.20.0
python-dotenv==0.21.0
